Clustering results
This is the first result of clustering. This is 1.5 months of data which contains 350000 samples and 65 features. 
The dimensions of the data are reduced and represented in the principal component space. Just 2 principle components explain 97.1% of the variance in the data. 

In the image, X-axis is principle component 1 and y-axis is principle component 2. 

Clustering is done through hdbscan algorithm and each color in the image represent one cluster. The black dots are data points which cannot be put into any cluster by the algorithm.